requirements.txt
beautifulsoup4==4.14.2
boto3==1.40.50
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.50.0
bedrock-agentcore
strands-agents
